simplify test
